Homework_Lesson38_Clouds_4_HM31/requirements.txt
google-cloud-firestore
google-cloud-compute